First commit of adding new file that will hold the code of Support
I added a new Directory containing new daily notes that will hold all of my notes from each commit. Along with this I created the first file that will hold all of the code to add support for the catalog based on the RestCatalog. I also added to the pom.xlm file with support for jdbc.